Misc edits (#47)
* cleanup

* Testing

* Query File

* Ignore vectors (they can be generated)

* Testing

* Data

* fixing rpc

* testing

* Misc edits, check to see if any should be kept

* master exec args array fix

* added tpc data, adding queueing back to master

* tpc dbms option

* removing python references

* range query messages

* updates, new data

* updates

* tring to fix memory issue

* working pointer issues

* slave address pointer null default

* pointer fix

* debug

* freeing different pointer

* debug

* debug

* debug

* slavelist updates, calling range query at proper time in slave

* consistent hashing fixes

* saved vector fix, committing fewer vectors, error reporting

* saved vector fix, committing fewer vectors, error reporting

* slave address fix, removing slavelist refs, debug

* removed superfluous comments, debugging send vector

* debug reallocation, vector counts

* updates

* updates

* ring consistent hashing fix

* proper return types for machine ids

* debug

* correction to rpc call

* slave death rpc

* timing experiment

* timing

* fault tolerance experimentation

* experiment updates

+++ types/types.h
@@ -4,9 +4,11 @@
 #ifndef TYPES_H
 #define TYPES_H
 #define MAX_VECTOR_LEN 128
+#include <sys/types.h>
+
 typedef struct vec_t {
-    unsigned long long vector[MAX_VECTOR_LEN];
-    unsigned int vector_length;
+    u_int64_t vector[MAX_VECTOR_LEN];
+    u_int vector_length;
 } vec_t;
-typedef unsigned int vec_id_t;
+typedef u_int vec_id_t;
 #endif /* TYPES_H */
